[Coverity] Analyze & Fix media-driver coverity VERY HIGH failures
Change:
- fixed overflowed constant and integer argument in media_softlet (INTEGER_OVERFLOW)

Tracked-On: NIACP3-1361

+++ media_softlet/agnostic/Xe_M_plus/Xe_LPM_plus_base/codec/hal/enc/hevc/features/encode_hevc_vdenc_const_settings_xe_lpm_plus_base.cpp
@@ -580,7 +580,7 @@ MOS_STATUS EncodeHevcVdencConstSettingsXe_Lpm_Plus_Base::SetVdencCmd2Settings()
                     data[32] |= (tmp1 << 16);
                     data[32] |= (tmp1 << 20);
                     data[32] |= (tmp0 << 24);
-                    data[32] |= (tmp0 << 28);
+                    data[32] |= ((uint32_t)tmp0 << 28);
 
                     data[33] |= tmp1;
                     data[33] |= (tmp1 << 4);
