+++ media_softlet/agnostic/common/codec/hal/enc/hevc/packet/encode_huc_brc_update_packet.cpp
@@ -1067,7 +1067,7 @@ namespace encode
 
     MHW_SETPAR_DECL_SRC(HCP_PIPE_MODE_SELECT, HucBrcUpdatePkt)
     {
-        params.codecStandardSelect = CodecHal_GetStandardFromMode(m_basicFeature->m_mode) - CODECHAL_HCP_BASE;
+        params.codecStandardSelect = (uint8_t)((int32_t)CodecHal_GetStandardFromMode(m_basicFeature->m_mode) - (int32_t)CODECHAL_HCP_BASE);
         params.bStreamOutEnabled   = true;
         params.bVdencEnabled       = true;
         params.multiEngineMode     = MHW_VDBOX_HCP_MULTI_ENGINE_MODE_FE_LEGACY;
